% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_ci.R
\name{fm_ci}
\alias{fm_ci}
\title{Farrington-Manning confidence interval for difference in proportions}
\usage{
fm_ci(df, m2, y, alpha = 0.025)
}
\arguments{
\item{df}{dataframe, output from sim_cont function}

\item{m2}{numeric, pre-specified fixed margin}

\item{y}{string, name of the outcome column}

\item{alpha}{numeric, one-sided alpha, Default: 0.025}
}
\value{
dataframe
}
\description{
calculates (1-alpha)100\% confidence interval for difference in
binomial proportions, and rejects/accepts the null hypothesis with respect
to the pre-specifid fixed margin
}
\details{
returns dataframe with 11 columns and 1 row, inlcuding estimated
proportion/n in each group, difference in proportions(group 'c' - group 't'),
variance of that difference (using Wald's method), lower and upper bounds of
the confidence interval and null hypothesis decision based on m2.
}
\examples{
 t <- sim_cont(
 p_C = 0.6, p_T = 0.5, n_arm = 200,
 mu1 = 4, mu2 = 100, sigma1 = 1,
 sigma2 = 20, r12 = -0.3,
 b1 = 0.1, b2 = -0.01)
 fm_ci(t, m2 = 0.1, 'y', alpha = 0.025)
}
\seealso{
\code{\link[rlang]{sym}}
 \code{\link[reshape2]{recast}}
 \code{\link[stats]{Normal}}
}
