% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi2comb.R
\name{mi2comb}
\alias{mi2comb}
\title{combintion following nested multiple imputation}
\usage{
mi2comb(dt, alpha = 0.025)
}
\arguments{
\item{dt}{dataframe}

\item{alpha}{numeric, one-sided alpha, Default: 0.025}
}
\value{
dataframe, final result following multiple imputation and
combination
}
\description{
combines data following nested multiple imputation
}
\seealso{
\code{\link[stats]{cor}},\code{\link[stats]{TDist}}
}
