% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi2comb_wn.R
\name{mi2comb_wn}
\alias{mi2comb_wn}
\title{combintion following nested multiple imputation for Wilson/Newcombe}
\usage{
mi2comb_wn(dt, alpha = 0.025)
}
\arguments{
\item{dt}{dataframe}

\item{alpha}{numeric, one-sided alpha, Default: 0.025}
}
\value{
dataframe, final result following multiple imputation and
combination
}
\description{
combines data following nested multiple imputation if Wilson/
Newcombe method is used for difference in binomial proportions
}
\seealso{
\code{\link[stats]{Normal}}
}
