% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_nonign_temp.R
\name{mice_nonign_temp}
\alias{mice_nonign_temp}
\title{Non-ingorably MI for difference in proportions via MICE}
\usage{
mice_nonign_temp(dt, n_mi = 5, m_mi = 10, m2, ci_method = ci_method,
  seed_mice, mu_C = 1, sd_C = 0, mu_T = 1, sd_T = 0)
}
\arguments{
\item{dt}{dataframe}

\item{n_mi}{numeric, number of patient level imputations, Default: 5}

\item{m_mi}{numeric, number of model imputations, Default: 10}

\item{m2}{numeric, margin}

\item{ci_method}{ci function name, Default: ci_method}

\item{seed_mice}{numeric}

\item{mu_C}{numeric, mean parameter for k multiplier normal distribution
for group 'c', Default: 1}

\item{sd_C}{numeric, stadnard deviation parameter for k multiplier normal
distribution for group 'c', Default: 0}

\item{mu_T}{Pnumeric, mean parameter for k multiplier normal distribution
for group 't',Default: 1}

\item{sd_T}{numeric, stadnard deviation parameter for k multiplier normal
distribution for group 't', Default: 0}
}
\value{
final result following nested multiple imputation
}
\description{
two-stage MI for difference in proportions using MICE
}
\seealso{
\code{\link[mice]{make.predictorMatrix}},\code{\link[mice]{mice}}
 \code{\link[rlang]{get_env}}
 \code{\link[tibble]{tibble}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[stats]{Uniform}},\code{\link[stats]{Normal}}
 \code{\link[tidyr]{unnest}},\code{\link[tidyr]{nest}}
}
