% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimpose.R
\name{mimpose}
\alias{mimpose}
\title{impose missing values}
\usage{
mimpose(dt, b_trt = 0, b_y = 0, b_x1 = 0, b_x2 = 0, b_ty = 0,
  do = 0.1, seed)
}
\arguments{
\item{dt}{dataframe, on which missing values are imposed}

\item{b_trt}{numeric, group missingness parameter, Default: 0}

\item{b_y}{numeric, outcome missingness parameter, Default: 0}

\item{b_x1}{numeric, covariate x1 missingness parameter, Default: 0}

\item{b_x2}{numeric, covariate x2 missingness parameter, Default: 0}

\item{b_ty}{numeric, covariatet treatment by group interaction, Default: 0}

\item{do}{numeric, target drop-out rate, Default: 0.1}

\item{seed}{numeric, seed specification before deleting observations}
}
\value{
original dataframe with missing outcome values
}
\description{
imposes missing values
}
\details{
this functtion returns the oroginal dataframe, where outcome
variable y is replaces with incomplete outcome y.m
}
\examples{
 t <- sim_cont(
 p_C = 0.6, p_T = 0.5, n_arm = 200,
 mu1 = 4, mu2 = 100, sigma1 = 1,
 sigma2 = 20, r12 = -0.3,
 b1 = 0.1, b2 = -0.01)
 tm <- mimpose(t, b_x1 = 0.2, seed = 1555)
}
\seealso{
\code{\link[stats]{Uniform}}
}
