% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_gen_an.R
\name{miss_gen_an}
\alias{miss_gen_an}
\title{generation and analysis of incomplete data}
\usage{
miss_gen_an(dt, m2, b_trt = 0, b_y = 0, b_x1 = 0, b_x2 = 0,
  b_ty = 0, do = 0.1, seed, dt_out = FALSE, sing_anal = TRUE,
  mice_anal = FALSE, bw_anal = FALSE, mu_C = 1, sd_C = 0,
  mu_T = 1, sd_T = 0, ci_method = wald_ci, seed_mice,
  alpha = 0.025, n_mi, m_mi, method = c("wald", "wn", "fm"),
  pro_wnmi = FALSE)
}
\arguments{
\item{dt}{dataframe, on which missing values are imposed}

\item{m2}{numeric, margin}

\item{b_trt}{numeric, group missingness parameter, Default: 0}

\item{b_y}{numeric, outcome missingness parameter, Default: 0}

\item{b_x1}{numeric, covariate x1 missingness parameter, Default: 0}

\item{b_x2}{numeric, covariate x2 missingness parameter, Default: 0}

\item{b_ty}{numeric, covariatet treatment by group interaction, Default: 0}

\item{do}{numeric, target drop-out rate, Default: 0.1}

\item{seed}{numeric, seed specification before deleting observations}

\item{dt_out}{logic, determines if the incomplete dataset needs to be
returned (if returned, no analysis will be done), Default: FALSE}

\item{sing_anal}{logic, determines if single value analysis needs to be
performed (complete case analysis, best and worst case scenario imputation),
Default: TRUE}

\item{mice_anal}{logic, determines if multiple imputation via MICE needs to
be performed (nested imputation), Default: FALSE}

\item{bw_anal}{logic, determines if the following single value imputation
needs to be performed: best case scenario for group 'c' and worst case
scenario for 't', Default: FALSE}

\item{mu_C}{numeric, mean parameter for k multiplier normal distribution
for group 'c', Default: 1}

\item{sd_C}{numeric, stadnard deviation parameter for k multiplier normal
distribution for group 'c', Default: 0}

\item{mu_T}{Pnumeric, mean parameter for k multiplier normal distribution
for group 't',Default: 1}

\item{sd_T}{numeric, stadnard deviation parameter for k multiplier normal
distribution for group 't', Default: 0}

\item{ci_method}{ci function name, Default: wald_ci}

\item{seed_mice}{numeric, Default: seed_mice}

\item{alpha}{numeric, one-sided alpha, Default: 0.025}

\item{n_mi}{numeric, number of multiple imputed patient level data}

\item{m_mi}{numeric, number of multiple imputed models}

\item{method}{character, method to construct confidence interval,
Default: c('wald','wn','fm')}

\item{pro_wnmi}{logic, specifies whether proper wn cobminaiton rules should
be used, if TRUE, then bin2mi package needs to be used, Default: FALSE}
}
\value{
list
}
\description{
imposes missing values, and performs analysis of incomplete data
}
\examples{
 t <- sim_cont(
 p_C = 0.6, p_T = 0.5, n_arm = 200,
 mu1 = 4, mu2 = 100, sigma1 = 1,
 sigma2 = 20, r12 = -0.3,
 b1 = 0.1, b2 = -0.01)
 miss_gen_an(t, m2 = 0.1 , b_x1 = 0.5, do = 0.1, seed = 1555, seed_mice = 1555, alpha = 0.025)
}
\seealso{
\code{\link[tidyr]{spread}}
 \code{\link[stats]{glm}}
 \code{\link[purrr]{set_names}}
}
