% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpars.R
\name{mpars}
\alias{mpars}
\title{Missingness parameter values}
\usage{
mpars(do = 0.2, atype)
}
\arguments{
\item{do}{numeric, Default: 0.2, drop out rate}

\item{atype}{string, analysis type: "sing" or mice}
}
\value{
tibble
}
\description{
assigns missingness parameter values
}
\details{
returns tibble with 6 columns: missingness type, and parameter
used in imposing missing data
}
\examples{
mpars(atype = "sing")
}
\seealso{
\code{\link[tibble]{tibble}}
}
